# docs/swagger.yaml
definitions:
  ncobase_common_resp.Exception:
    properties:
      code:
        description: Business code
        type: integer
      data:
        description: Response data
      errors:
        description: Validation errors
      message:
        description: Message
        type: string
      status:
        description: HTTP status
        type: integer
    type: object
  ncobase_common_types.JSON:
    additionalProperties: {}
    type: object
...
  structs.CreateModuleBody:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      markdown:
        type: boolean
      name:
        type: string
      private:
        type: boolean
      released:
        description: Use pointer for nullable field
        type: string
      slug:
        type: string
      status:
        description: Use pointer for nullable field
        type: integer
      temp:
        type: boolean
      thumbnail:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  structs.ReadModule:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      extras:
        $ref: '#/definitions/ncobase_common_types.JSON'
      id:
        type: string
      keywords:
        items:
          type: string
        type: array
      markdown:
        type: boolean
      private:
        type: boolean
      released:
        type: string
      slug:
        type: string
      status:
        type: integer
      temp:
        type: boolean
      thumbnail:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  structs.UpdateModuleBody:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      markdown:
        type: boolean
      name:
        type: string
      private:
        type: boolean
      released:
        description: Use pointer for nullable field
        type: string
      slug:
        type: string
      status:
        description: Use pointer for nullable field
        type: integer
      temp:
        type: boolean
      thumbnail:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
  description: Next-Gen Modular Framework
  title: Ncobase
  version: 0.1.0
paths:
...
  /v1/modules:
    get:
      description: Retrieve a list of modules based on the provided query parameters
      parameters:
      - in: query
        name: cursor
        type: string
      - in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/structs.ReadModule'
            type: array
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: List all modules
      tags:
      - modules
    post:
      consumes:
      - application/json
      description: Create a new module with the provided data
      parameters:
      - description: Module data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.CreateModuleBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadModule'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Create a new module
      tags:
      - modules
  /v1/modules/{slug}:
    delete:
      description: Delete a module by its slug or ID
      parameters:
      - description: Module slug or ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadModule'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Delete a module by slug or ID
      tags:
      - modules
    get:
      description: Retrieve a module by its slug or ID
      parameters:
      - description: Module slug or ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadModule'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get a module by slug or ID
      tags:
      - modules
    put:
      consumes:
      - application/json
      description: Update an existing module with the provided data
      parameters:
      - description: Module slug or ID
        in: path
        name: slug
        required: true
        type: string
      - description: Module data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateModuleBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadModule'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Update an existing module
      tags:
      - modules
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"